Test"Battle_Ground" for theme oop-principles.
